FastAPI_BE/requirements.txt
# FastAPI Backend Requirements

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ML libraries
transformers>=4.45.0
accelerate>=0.34.0
safetensors>=0.4.0
torch>=2.4.0
torchvision>=0.19.0

# Image processing
pillow>=10.0.0

# Utilities
pydantic>=2.0.0
